fix(em): preserve source Euler metadata through pose updates
(cherry picked from commit df9975ee2c11f749f625442fecdfc7ab7bd1d153)

+++ recovar/em/dense_single_volume/iteration_loop.py
@@ -4030,7 +4030,7 @@ def _run_relion_iteration_loop(
             if best_pose_rotations[k] is not None:
                 best_rots = np.asarray(best_pose_rotations[k], dtype=_pose_state_dtype)
                 best_eulers = (
-                    np.asarray(best_pose_rotation_eulers[k], dtype=_pose_state_dtype)
+                    np.asarray(best_pose_rotation_eulers[k], dtype=np.float64)
                     if best_pose_rotation_eulers[k] is not None
                     else utils.R_to_relion(best_rots, degrees=True).astype(_pose_state_dtype)
                 )
